backend/requirements.txt
django>=5.0,<6.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.3,<5.0
channels>=4.0,<5.0
daphne>=4.0,<5.0
psycopg2-binary>=2.9,<3.0
python-decouple>=3.8,<4.0
 
# Uncomment for production WebSocket scaling (requires Redis)
# channels-redis>=4.0,<5.0
 